Main

See merge request tresearchgroup/programming-department/projects/palila/alpha/framework!1

+++ src/main/java/tech/tresearchgroup/palila/controller/search/GenericSAO.java
@@ -6,6 +6,8 @@ import com.meilisearch.sdk.SearchRequest;
 import com.meilisearch.sdk.Task;
 import com.meilisearch.sdk.model.SearchResult;
 import tech.tresearchgroup.palila.controller.database.GenericDAO;
+import tech.tresearchgroup.palila.model.BaseSettings;
+import tech.tresearchgroup.palila.model.enums.SearchMethodEnum;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -28,33 +30,43 @@ public class GenericSAO implements GenericSearchFunctionality {
 
     @Override
     public void createDocument(Object object, Index index) throws Exception {
-        index.addDocuments(gson.toJson(object));
+        if (BaseSettings.searchMethod.equals(SearchMethodEnum.SEARCH_SERVER)) {
+            index.addDocuments(gson.toJson(object));
+        }
     }
 
     @Override
     public void updateDocument(Object object, Index index) throws Exception {
-        index.updateDocuments(gson.toJson(object));
+        if (BaseSettings.searchMethod.equals(SearchMethodEnum.SEARCH_SERVER)) {
+            index.updateDocuments(gson.toJson(object));
+        }
     }
 
     @Override
     public void deleteDocument(long id, Index index) throws Exception {
-        index.deleteDocument(String.valueOf(id));
+        if (BaseSettings.searchMethod.equals(SearchMethodEnum.SEARCH_SERVER)) {
+            index.deleteDocument(String.valueOf(id));
+        }
     }
 
     @Override
     public void deleteAllDocuments(Index index) throws Exception {
-        Task task = index.deleteAllDocuments();
-        task.wait();
+        if (BaseSettings.searchMethod.equals(SearchMethodEnum.SEARCH_SERVER)) {
+            Task task = index.deleteAllDocuments();
+            task.wait();
+        }
     }
 
     @Override
     public void reindex(int maxResultsSize, GenericDAO genericDAO, Index index, Class theClass) throws Exception {
-        long totalPages = genericDAO.getTotalPages(maxResultsSize, theClass);
-        for (int i = 0; i <= totalPages; i++) {
-            List items = genericDAO.readPaginated(maxResultsSize, i, theClass);
-            for (Object item : items) {
-                String json = gson.toJson(item);
-                index.addDocuments(json);
+        if (BaseSettings.searchMethod.equals(SearchMethodEnum.SEARCH_SERVER)) {
+            long totalPages = genericDAO.getTotalPages(maxResultsSize, theClass);
+            for (int i = 0; i <= totalPages; i++) {
+                List items = genericDAO.readPaginated(maxResultsSize, i, theClass, false);
+                for (Object item : items) {
+                    String json = gson.toJson(item);
+                    index.addDocuments(json);
+                }
             }
         }
     }
